Struct er_save_lib::Save

pub struct Save {
    pub magic: [u8; 4],
    /* private fields */
}

A struct representing an Elden Ring save file.

This struct provides methods to read a save file from a specified file path or from a byte slice, and to check if a byte slice can be parsed as a valid save file. It includes functions for reading, parsing, and validating save files, ensuring that implementations handle the respective file formats correctly.

Fields
magic: [u8; 4]
Implementations

impl Save

pub fn write_to_vec(&self) -> Result<Vec<u8>, DekuError>

Writes the save file to a byte vector and returns it as a Vec<u8>.

This function converts the Save instance into a byte vector, which can then be saved to a file or used as needed.

Errors

This function will return an Err variant if any of the following conditions occur:

There are issues with formatting the data into a byte slice.

Possible error types include:

deku::DekuError if the bytes cannot be formatted correctly.

Examples

use er_save_lib::Save;

fn main() {
    let save = Save::from_path("./test/ER0000.sl2").expect("Failed to read save file!");
    let bytes = save.write_to_vec().expect("Failed to write save file!");
}

Safety

This function is safe to call as it only performs data formatting operations.

pub fn write_to_path( &self, path: impl AsRef<Path>, ) -> Result<(), SaveParseError>

Writes the save file to a specified path.

This function attempts to write the Save instance to a file at the given path.

Errors

This function will return an Err variant if any of the following conditions occur:

The file cannot be created or opened for writing.

There are issues with writing the data to the file.

Possible error types include:

std::io::Error if there are issues with file access.

deku::DekuError if the bytes cannot be formatted correctly.

Examples

use er_save_lib::Save;

fn main() {
    let save = Save::from_path("./test/PS_Save.txt").expect("Failed to read save file!");
    save.write_to_path("./test/New_Save.txt").expect("Failed to write save file!");
}

Safety

This function is safe to call as it only performs file writing and data formatting operations.

pub fn from_path(path: impl AsRef<Path>) -> Result<Self, SaveParseError>

Reads a save file from the specified path and returns a Save instance.

This function attempts to read the file at the given path, parse its contents as a Save structure, and return it wrapped in a Result.

Errors

This function will return an Err variant if any of the following conditions occur:

The file does not exist at the specified path.

There are insufficient permissions to read the file.

The file is not valid or cannot be parsed as a Save structure.

Possible error types include:

std::io::Error if there are issues with file access.

deku::DekuError if the bytes cannot be parsed correctly.

Examples

use er_save_lib::Save;

fn main() {
    let save = Save::from_path("./test/PS_Save.txt").expect("Failed to read save file!");
}

Panics

This function does not explicitly panic. However, if it is called in a context where the Result is not handled properly, it could cause a panic.

Safety

This function is safe to call as it only performs file reading and parsing operations.

pub fn from_slice(bytes: &[u8]) -> Result<Self, SaveParseError>

Parses a byte slice and returns a Save instance.

This function attempts to parse the provided byte slice as a Save structure and return it wrapped in a Result.

Errors

std::io::Error if there are issues with file access.

deku::DekuError if the bytes cannot be parsed correctly.

This function will return an Err variant if the byte slice cannot be parsed as a valid Save structure.

Possible error types include:

deku::DekuError if the bytes cannot be parsed correctly.

Examples

use er_save_lib::Save;

fn main() {
    let bytes = std::fs::read("./test/PS_Save.txt").unwrap();
    let save = Save::from_slice(&bytes).expect("Failed to read save file!");
}

Panics

This function does not explicitly panic. However, if it is called in a context where the Result is not handled properly, it could cause a panic.

Safety

This function is safe to call as it only performs parsing operations on the provided byte slice.

pub fn is(bytes: &[u8]) -> bool

Checks if the provided byte slice can be parsed as a valid Save structure.

This function performs a quick validation check on the byte slice to determine if it can be parsed as a Save structure. This can be useful for filtering or validating data before attempting to fully parse it.

Examples

use er_save_lib::Save;

fn main() {
    let bytes = std::fs::read("./test/ER0000.sl2").unwrap();
    assert!(Save::is(&bytes));
}

Safety

This function is safe to call as it only performs validation operations on the provided byte slice.

pub fn is_ps(bytes: &[u8]) -> bool

Checks if the provided byte slice indicates a PlayStation save file.

This function performs a quick validation check on the byte slice to determine if it belongs to a PlayStation save file format. This can be useful for filtering or validating data before attempting to fully parse it.

Examples

use er_save_lib::Save;

fn main() {
    let bytes = std::fs::read("./test/PS_Save.txt").unwrap();
    assert!(Save::is_ps(&bytes));
}

Safety

This function is safe to call as it only performs validation operations on the provided byte slice.

pub fn is_pc(bytes: &[u8]) -> bool

Checks if the provided byte slice indicates a PC save file.

This function performs a quick validation check on the byte slice to determine if it belongs to a PC save file format. This can be useful for filtering or validating data before attempting to fully parse it.

Examples

use er_save_lib::Save;

fn main() {
    let bytes = std::fs::read("./test/ER0000.sl2").unwrap();
    assert!(Save::is_pc(&bytes));
}

Safety

This function is safe to call as it only performs validation operations on the provided byte slice.