Struct er_save_lib::SaveApi

pub struct SaveApi { /* private fields */ }

Implementations

impl SaveApi

pub fn to_vec(&self) -> Result<Vec<u8>, SaveApiError>

Converts the save data to a vector of bytes.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let bytes = save_api.to_vec().unwrap();

pub fn write_to_path(&self, path: impl AsRef<Path>) -> Result<(), SaveApiError>

Writes the save data to the specified path.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.write_to_path("./test/null.sl2").unwrap();

pub fn platform(&self) -> SaveType

Returns the platform type of the save file.

Example

use er_save_lib::{SaveApi, SaveType};
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let platform = save_api.platform();
assert_eq!(platform, SaveType::PC);

pub fn steam_id(&self) -> u64

Returns the Steam ID associated with the save file.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let steam_id = save_api.steam_id();

pub fn set_steam_id(&mut self, steam_id: u64) -> Result<(), SaveApiError>

Sets the Steam ID associated with the save file.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_steam_id(1234567890).unwrap();

impl SaveApi

pub fn character_index_from_name(&self, name: &str) -> Option<usize>

Returns the index of the character with the given name.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let index = save_api.character_index_from_name("CharacterName");

pub fn active_characters(&self) -> [bool; 10]

Returns an array indicating which characters are active.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let active_characters = save_api.active_characters();

impl SaveApi

pub fn get_param<P: Param>(&self) -> Result<Param<P>, SaveApiError>

Returns parameters by type

Example

use er_save_lib::SaveApi;
use er_save_lib::TalkParam::TalkParam;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let param = save_api.get_param::<TalkParam>().unwrap();

pub fn get_param_bytes_map( &self, ) -> Result<&BTreeMap<String, Vec<u8>>, SaveApiError>

Returns a map of parameter bytes.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let param_bytes_map = save_api.get_param_bytes_map().unwrap();

impl SaveApi

pub fn set_archetype( &mut self, index: usize, archetype: u8, ) -> Result<(), SaveApiError>

Sets the archetype of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let (index, archetype_id) = (0, 1);
save_api.set_archetype(index, archetype_id);

pub fn set_level( &mut self, index: usize, level: u32, ) -> Result<(), SaveApiError>

Sets the level of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_level(0, 1);

pub fn set_runes_memory( &mut self, index: usize, runes_memory: u32, ) -> Result<(), SaveApiError>

Sets the rune memory of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_runes_memory(0, 1_000);

pub fn set_character_name( &mut self, index: usize, new_name: &str, ) -> Result<(), SaveApiError>

Sets the name of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_character_name(0, "NewName").unwrap();

pub fn set_gender( &mut self, index: usize, gender: u8, ) -> Result<(), SaveApiError>

Sets the gender of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let (index, gender_id) = (0, 1);
save_api.set_gender(index, gender_id);

impl SaveApi

pub fn hp(&self, index: usize) -> u32

Returns the hp for the character at the specified index.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let hp = save_api.hp(0);

pub fn equipped_gestures(&self, index: usize) -> &Vec<u32>

Returns the equipped gestures for the character at the specified index.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let equipped_gestures = save_api.equipped_gestures(0);

pub fn set_equipped_gestures( &mut self, index: usize, new_gestures: Vec<u32>, ) -> Result<(), SaveApiError>

Sets the equipped gestures for the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_equipped_gestures(0, vec![1u32,2u32,3u32]).unwrap();

pub fn set_hp(&mut self, index: usize, hp: u32) -> Result<(), SaveApiError>

Sets the hp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let (index, hp) = (0, 1_000);
save_api.set_hp(index, hp);

pub fn max_hp(&self, index: usize) -> u32

Gets the max hp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let index = 0;
save_api.max_hp(index);

pub fn set_max_hp( &mut self, index: usize, max_hp: u32, ) -> Result<(), SaveApiError>

Sets the max hp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let (index, max_hp) = (0, 1_000);
save_api.set_max_hp(index, max_hp);

pub fn base_max_hp(&self, index: usize) -> u32

Gets the base max hp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let index = 0;
save_api.base_max_hp(index);

pub fn set_base_max_hp( &mut self, index: usize, base_max_hp: u32, ) -> Result<(), SaveApiError>

Sets the base max hp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let (index, base_max_hp) = (0, 1_000);
save_api.set_base_max_hp(index, base_max_hp);

pub fn set_fp(&mut self, index: usize, fp: u32) -> Result<(), SaveApiError>

Sets the fp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_fp(0, 1);

pub fn max_fp(&self, index: usize) -> u32

Gets the max fp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.max_fp(0);

pub fn set_max_fp( &mut self, index: usize, max_fp: u32, ) -> Result<(), SaveApiError>

Sets the max fp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_max_fp(0, 1);

pub fn base_max_fp(&self, index: usize) -> u32

Gets the base max fp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.base_max_fp(0);

pub fn set_base_max_fp( &mut self, index: usize, base_max_fp: u32, ) -> Result<(), SaveApiError>

Sets the base max fp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_base_max_fp(0, 1);

pub fn set_sp(&mut self, index: usize, sp: u32) -> Result<(), SaveApiError>

Sets the sp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_sp(0, 1);

pub fn max_sp(&self, index: usize) -> u32

Gets the max sp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.max_sp(0);

pub fn set_max_sp( &mut self, index: usize, max_sp: u32, ) -> Result<(), SaveApiError>

Sets the max sp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_max_sp(0, 1);

pub fn base_max_sp(&self, index: usize) -> u32

Gets the base max sp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.base_max_sp(0);

pub fn set_base_max_sp( &mut self, index: usize, base_max_sp: u32, ) -> Result<(), SaveApiError>

Sets the base max sp of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_base_max_sp(0, 1);

pub fn level(&self, index: usize) -> u32

Gets the level of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.level(0);

pub fn vigor(&self, index: usize) -> u32

Gets the vigor of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.vigor(0);

pub fn set_vigor( &mut self, index: usize, vigor: u32, ) -> Result<(), SaveApiError>

Sets the vigor of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_vigor(0, 1);

pub fn mind(&self, index: usize) -> u32

Gets the mind of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.mind(0);

pub fn set_mind(&mut self, index: usize, mind: u32) -> Result<(), SaveApiError>

Sets the mind of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_mind(0, 1);

pub fn endurance(&self, index: usize) -> u32

Gets the endurance of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.endurance(0);

pub fn set_endurance( &mut self, index: usize, endurance: u32, ) -> Result<(), SaveApiError>

Sets the endurance of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_endurance(0, 1);

pub fn strength(&self, index: usize) -> u32

Gets the strength of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.strength(0);

pub fn set_strength( &mut self, index: usize, strength: u32, ) -> Result<(), SaveApiError>

Sets the strength of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_strength(0, 1);

pub fn dexterity(&self, index: usize) -> u32

Gets the dexterity of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.dexterity(0);

pub fn set_dexterity( &mut self, index: usize, dexterity: u32, ) -> Result<(), SaveApiError>

Sets the dexterity of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_dexterity(0, 1);

pub fn intelligence(&self, index: usize) -> u32

Gets the intelligence of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.intelligence(0);

pub fn set_intelligence( &mut self, index: usize, intelligence: u32, ) -> Result<(), SaveApiError>

Sets the intelligence of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_intelligence(0, 1);

pub fn faith(&self, index: usize) -> u32

Gets the faith of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.faith(0);

pub fn set_faith( &mut self, index: usize, faith: u32, ) -> Result<(), SaveApiError>

Sets the faith of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_faith(0, 1);

pub fn arcane(&self, index: usize) -> u32

Gets the arcane of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.arcane(0);

pub fn set_arcane( &mut self, index: usize, arcane: u32, ) -> Result<(), SaveApiError>

Gets the arcane of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_arcane(0, 1);

pub fn runes(&self, index: usize) -> u32

Gets the runes of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.runes(0);

pub fn set_runes( &mut self, index: usize, runes: u32, ) -> Result<(), SaveApiError>

Sets the runes of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_runes(0, 1_000);

pub fn runes_memory(&self, index: usize) -> u32

Gets the rune memory of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.runes_memory(0);

pub fn regions(&self, index: usize) -> Result<&Vec<u32>, SaveApiError>

Gets the regions of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.regions(0);

pub fn regions_count(&self, index: usize) -> Result<u32, SaveApiError>

Gets the regions count of the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.regions_count(0);

pub fn add_region( &mut self, index: usize, region_id: u32, ) -> Result<(), SaveApiError>

Adds a region to the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.add_region(0, 1_000);

pub fn remove_region( &mut self, index: usize, region_id: u32, ) -> Result<(), SaveApiError>

Removes a region from the character at the specified index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.remove_region(0, 1_000);

pub fn archetype(&self, index: usize) -> u8

Returns the id of the archetype for the character at the specified index.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let archetype = save_api.archetype(0);

pub fn gender(&self, index: usize) -> u8

Returns the gender of the character at the specified index.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let gender = save_api.gender(0);

pub fn character_name(&self, index: usize) -> String

Returns the name of the character at the specified index.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
let name = save_api.character_name(0);

impl SaveApi

pub fn new(save: Save) -> Self

Creates a new SaveApi instance from a Save object.

Example

use er_save_lib::SaveApi;
use er_save_lib::Save;
let save = Save::from_path("./test/PS_Save.txt").unwrap();
let save_api = SaveApi::new(save);

pub fn from_slice(bytes: &[u8]) -> Result<Self, SaveApiError>

Creates a SaveApi instance from a slice of bytes.

Example

use er_save_lib::SaveApi;
let bytes = std::fs::read("./test/ER0000.sl2").unwrap();
let save_api = SaveApi::from_slice(&bytes).unwrap();

pub fn from_path(path: impl AsRef<Path>) -> Result<Self, SaveApiError>

Creates a SaveApi instance from a file path.

Example

use er_save_lib::SaveApi;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();

impl SaveApi

pub fn get_event_flag( &self, event_id: u32, character_index: usize, ) -> Result<bool, SaveApiError>

Returns event flags by id and character index

Example

use er_save_lib::SaveApi;
use er_save_lib::TalkParam::TalkParam;
let save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
// IDs can be found in src\res\eventflag_bst.txt as the second element of each tuple
let event_flag = save_api.get_event_flag(6223, 0).unwrap();

pub fn set_event_flag( &mut self, event_id: u32, character_index: usize, on: bool, ) -> Result<(), SaveApiError>

Sets the value of the specified event flag for the given character index.

Example

use er_save_lib::SaveApi;
let mut save_api = SaveApi::from_path("./test/ER0000.sl2").unwrap();
save_api.set_event_flag(123, 0, true).unwrap();